Add defaults when ruleId is null
Sometimes ESLint does not always provide a ruleId in the message
objects. This can happen if, for example, you have your ecmaVersion
set to 5 but are using ES6+ features. ESLint would fail to parse
the syntax and return a fatal error. A ruleId would not be able
to be assigned as it would be unable to match on anything.

This is problematic, especially for the inspections formatter as
TeamCity complains if an inspection is returned without an id or
typeId.

This PR addresses this by using a default of '<none>' of no ruleId
is found.